Updated shared_preprocess to resolve CT FIPS error & added 2022 ACS populations, and the file can now be run through the CLI. Still working through the superfund scoring script! It works locally for one state but need to do some additional testing.

+++ scripts/shared/shared_preprocess.R
@@ -7,11 +7,11 @@
 # to a given block. This table is often seen in EJScreen data processing script 
 # to weight datasets. 
 # 
-# TODO - translate this code to match the pipeline 
-# TODO - create args for code for: local vs remote, location (req), and state (opt) 
-# TODO - keep national file push 
-# TODO - add data for the territories 
-# 
+# TODO - code currently uses ACS 2022 vintage as the null-override score, should
+#        we make this an argument?
+# TODO - this currently writes to version 1.0, should we make version an 
+#        argument as well?
+#
 # Outputs: all from 2020 decennial census 
 #   - state_abb: abbreviated state name 
 #   - STATEFP20: state number 
@@ -34,11 +34,33 @@
 #       I can confirm the spatial boundaries did not change, it was just simply 
 #       a swap in census codes. 
 # 
+# Example uses in CLI:
+#     # for local testing: 
+#     Rscript scripts/shared/shared_preprocess.R -l local
+#     # for adding it to s3: 
+#     Rscript scripts/shared/shared_preprocess.R -l remote
+# 
 # Date: Feb 24th, 2026 
-# Updated: June 3rd, 2026
+# Updated: July 28th, 2026
 # Author: EmmaLi Tsai
 ###############################################################################
+# package handling 
+###############################################################################
+suppressPackageStartupMessages({
+  required <- c('tigris','janitor','tidyverse','curl','tidycensus','data.table')
+  for (pkg in required) {
+    if (!requireNamespace(pkg, quietly = TRUE)) {
+      stop(sprintf('Required package "%s" is not installed. Install it and retry.', 
+                   pkg))
+    }
+  }
+  # aws.s3 is optional and only used if s3 URIs are provided
+  has_aws <- requireNamespace('aws.s3', quietly = TRUE)
+})
+
+###############################################################################
 # libraries 
+###############################################################################
 library(tigris)
 library(janitor)
 library(tidyverse)
@@ -47,12 +69,34 @@ library(sf)
 library(tidycensus)
 library(aws.s3)
 library(data.table)
+library(optparse)
 # cache tigris files:
 options(tigris_use_cache = TRUE)
 options(scipen = 9999)
 
+###############################################################################
+# constants 
+###############################################################################
 # setting ACS year for block group populations
-acs_year = 2022
+ACS_YEAR = 2022
+
+###############################################################################
+# Handling arguments - really simple location & state argument 
+###############################################################################
+option_list <- list(
+  make_option(c('-l','--location'), type='character', default='local', 
+              help='Select location: local or remote (default: local)'))
+  # removing the state option, as we decided to move forward with a national 
+  #file
+  # make_option(c('-s','--state'), type='character', default=NULL, 
+  #             help="Required two-letter state code or 'all'"))
+
+parser <- OptionParser(usage='%prog [options]', option_list=option_list)
+args_all <- parse_args(parser, positional_arguments = FALSE)
+opts <- args_all
+
+###############################################################################
+# Handling states 
 ###############################################################################
 # pull state IDs and a clean acronym for looping: 
 state_acro <- states() %>%
@@ -60,10 +104,19 @@ state_acro <- states() %>%
   janitor::clean_names() %>%
   arrange(stusps) %>%
   select(statefp, stusps) %>%
-  # TODO - there is definitely a way to collect populations for these territories, 
-  # I just can't figure it out right now 
+  # we don't be able to get weights from the territories, unfortunately
   filter(!(stusps %in% c("VI", "MP", "GU", "AS")))
 
+# removing the state option, as we decided to move forward with a national 
+# if(opts$state != "all") {
+#   # filter if state option is not all 
+#   state_acro <- state_acro %>%
+#     filter(stusps == opts$state)
+# } 
+
+###############################################################################
+# Looping over states
+###############################################################################
 # loop through states
 weights_df <- data.frame() 
 for (i in 1:nrow(state_acro)) {
@@ -88,7 +141,6 @@ for (i in 1:nrow(state_acro)) {
   state_i_b <- st_read(paste0(file_loc, "/tl_2020_", state_num_i, "_tabblock20.shp")) %>%
     # create the block group geoid
     mutate(block_group_geoid = substr(GEOID20, 0, 12))
-  
   
   # group decennial block data by census block group geoid, and sun 
   # populations to create the weights table
@@ -121,7 +173,7 @@ for (i in 1:nrow(state_acro)) {
     geography = "block group", 
     variables = "B01003_001", 
     state = state_num_i,
-    year = as.numeric(acs_year),
+    year = as.numeric(ACS_YEAR),
     geometry = F) %>%
     rename(block_group_geoid_2022 = GEOID, 
            acs_2022_bg_pop = estimate) %>%
@@ -138,7 +190,7 @@ for (i in 1:nrow(state_acro)) {
   # Based on my understanding & considering the actual bounaries didn't change, 
   # we can remap them pretty easily: 
   if(length(unique(state_i_bg_2022$block_group_geoid_2022)) != length(unique(b_weights_i$block_group_geoid))){
-    print("Fixing FIPS in CT")
+    print(paste0("Fixing FIPS in: ", state_i))
     # reading in CT crosswalk
     ct_xwalk <- data.table::fread("https://raw.githubusercontent.com/CT-Data-Collaborative/2022-block-crosswalk/main/2022blockcrosswalk.csv")
     # tidying and cleaning fips codes: 
@@ -152,6 +204,7 @@ for (i in 1:nrow(state_acro)) {
              block_group_2022 = substr(block_fips_2022, 1, 12)) %>%
       select(block_group_2020, block_group_2022) 
     
+    # NA here for 2022 bg: 091909900000
     state_i_bg_2022_ct <- merge(state_i_bg_2022, 
                                 ct_xwalk_tidy, 
                                 by.x = "block_group_geoid_2022", 
@@ -162,7 +215,14 @@ for (i in 1:nrow(state_acro)) {
                                 state_i_bg_2022_ct, 
                                 by.x = "block_group_geoid", 
                                 by.y = "block_group_2020",
-                                all = T)
+                                all = T) %>%
+      # NAN fraction of totals == 0/0
+      # missing 2022 bg = mostly water
+      # missing 2020 bg = zero pops in 2020, but in 2022
+      mutate(fraction_of_total = case_when(is.na(fraction_of_total) ~ 0, 
+                                           TRUE ~ fraction_of_total), 
+             acs_2022_bg_pop = case_when(is.na(acs_2022_bg_pop) ~ 0, 
+                                         TRUE ~ acs_2022_bg_pop))
     
   } else {
     # mergin: 
@@ -175,32 +235,47 @@ for (i in 1:nrow(state_acro)) {
     
   }
   
-
   
   # binding 
   weights_df <- bind_rows(weights_df, b_weights_i_merged)
 }
 
-# Trimming any extra duplicates 
+# trimming any extra duplicates 
 weights_df_uniq <- weights_df %>%
   unique() %>%
   select(-VINTAGE)
 
-# adding to s3: 
-tmp <- tempfile()
-write.csv(weights_df_uniq, paste0(tmp, ".csv"), row.names = F)
-on.exit(unlink(tmp))
-put_object(
-  file = paste0(tmp, ".csv"),
-  object = "s3://pedp-data-preserved/ejscreen-data-processing/census_tables/census_block_weights_2020_w2022pops.csv",
-  multipart = T
-)
+# if option is local, save to shared/pipeline folder 
+if(opts$location == "local"){
+  print(paste0("Saving Data Locally"))
+  script_dir <- getwd()
+  save_path <- normalizePath(file.path(script_dir, 'pipeline/shared/census_block_weights/1.0/preprocessed_input/census_block_weights_2020/'))
+  save_path_name <- paste0(save_path, "/census_block_weights_2020_w2022pops.csv")
+  write.csv(weights_df_uniq, save_path_name)
+  
+} else {
+  print(paste0("Saving Data to S3"))
+  
+  # if not, save to s3 
+  save_path_name <- paste0("s3://pedp-data-preserved/ejscreen-data-processing/pipeline/shared/census_block_weights/1.0/preprocessed_input/census_block_weights_2020_w2022pops.csv")
+  
+  tmp <- tempfile()
+  write.csv(weights_df_uniq, paste0(tmp, ".csv"), row.names = F)
+  on.exit(unlink(tmp))
+  put_object(
+    file = paste0(tmp, ".csv"),
+    object = save_path_name,
+    multipart = T
+  )
+}
 
 
-### extra code to check output against previous file in s3: 
+###############################################################################
+# extra code to check output against previous file in s3: 
+###############################################################################
 # # looking at differences against our original file: 
-# test <- aws.s3::s3read_using(read.csv, 
-#                              object = "s3://pedp-data-preserved/ejscreen-data-processing/census_tables/census_block_weights_2020.csv")
+# test <- aws.s3::s3read_using(read.csv,
+#                              object = "s3://pedp-data-preserved/ejscreen-data-processing/census_tables/census_block_weights_2020_w2022pops.csv")
 # # fixing geoids & leading 0s
 # test_tidy <- test %>%
 #   mutate(GEOID20 = as.character(GEOID20),
